+++ src/Console/Commands/DeployCommand.php
@@ -2,22 +2,15 @@
 
 namespace WebId\Radis\Console\Commands;
 
-use WebId\Radis\Console\Commands\Traits\CheckGitBranch;
-use WebId\Radis\Console\Commands\Traits\HasStub;
-
 class DeployCommand extends ForgeAbstractCommand
 {
-    use HasStub,
-        CheckGitBranch;
+    /** @var string  */
+    protected $signature = 'radis:update
+                            {site_name : Site name on forge}
+                            {--site=} : Site ID on forge';
 
     /** @var string  */
-    protected $signature = 'radis:deploy
-                            {site_name : Name to set on forge}
-                            {git_branch : Name of the git branch to deploy}
-                            {--database=} : Database name on forge';
-
-    /** @var string  */
-    protected $description = 'Deploy a Review App';
+    protected $description = 'Deploy existing Review App';
 
     /**
      * Execute the console command.
@@ -26,53 +19,18 @@ class DeployCommand extends ForgeAbstractCommand
      */
     public function handle()
     {
-        $this->checkConfig('radis.forge.digital_ocean_api_key');
-        $this->checkConfig('radis.git_repository');
-
         $siteName = $this->argument('site_name');
-        $gitBranch = $this->argument('git_branch');
-        if (!$this->checkGitBranch($gitBranch)) {
+
+        $siteId = $this->option('site');
+
+        $site = $this->getSite($siteName, $siteId);
+        if (!$site) {
             return 0;
         }
-        $databaseName = $this->option('database');
-
-        $featureDomain = $this->forgeService->getFeatureDomain($siteName);
-
-        $this->destroyExisting($siteName, $databaseName);
-
-        $this->info('Creating forge site : "'.$featureDomain.'"...');
-        $site = $this->forgeService->createForgeSite($this->forgeServer, $siteName, $gitBranch, $databaseName);
-
-        $this->callSilent('radis:env', [
-            'site_name' => $siteName,
-            '--site' => $site->id
-        ]);
-
-        $this->callSilent('radis:deploy-script', [
-            'site_name' => $siteName,
-            'git_branch' => $gitBranch,
-            '--site' => $site->id
-        ]);
 
         $site->deploySite();
 
-        $this->info("The review app `${siteName}` will be created with the branch `${gitBranch}`");
+        $this->info("The review app `${siteName}` will be deployed");
 
-        return 0;
-    }
-
-    /**
-     * @param string $siteName
-     * @param string|null $databaseName
-     */
-    private function destroyExisting(string $siteName, string $databaseName = null)
-    {
-        $commandDestroy = [
-            'site_name' => $siteName,
-        ];
-        if ($databaseName) {
-            $commandDestroy['--database'] = $databaseName;
-        }
-        $this->call('radis:destroy', $commandDestroy);
     }
 }
